refactor(mgmt): move run_manager_backup to dedicated backup module to avoid circular imports
Extracted the `run_manager_backup` function from `sdcm/mgmt/operations.py` into a new
dedicated module `sdcm/mgmt/backup.py`. This breaks the circular import dependency chain:
`sdcm/tester.py` → `sdcm/nemesis.py` → `sdcm/mgmt/operations.py` → `sdcm/tester.py`.

The backup module only imports from non-tester modules (`sdcm.mgmt`, `sdcm.sct_events`),
eliminating the circular dependency while maintaining the same functionality.

Updated import in `sdcm/nemesis.py` to import from the new backup module.

Changes:
- Created `sdcm/mgmt/backup.py` with `run_manager_backup` function
- Removed `run_manager_backup` from `sdcm/mgmt/operations.py`
- Updated import in `sdcm/nemesis.py`

This resolves the ImportError that prevented unit tests from loading when importing
ClusterTester, which depends on nemesis which was trying to import from operations.
